+++ src/main/java/com/epam/musicbox/repository/ArtistRepository.java
@@ -10,9 +10,9 @@ import java.util.Optional;
 public interface ArtistRepository extends Repository<Artist> {
     Optional<Artist> findByName(String name);
 
-    List<Track> getTracks(Integer artistId, int offset, int limit) throws HttpException;
+    List<Track> getTracks(Long artistId, int offset, int limit) throws HttpException;
 
-    void addTrack(Integer playlistId, Integer trackId) throws HttpException;
+    void addTrack(Long playlistId, Long trackId) throws HttpException;
 
-    void removeTrack(Integer playlistId, Integer trackId) throws HttpException;
+    void removeTrack(Long playlistId, Long trackId) throws HttpException;
 }
