+++ src/main/java/com/epam/musicbox/repository/UserRepository.java
@@ -11,33 +11,33 @@ public interface UserRepository extends Repository<User> {
 
     Optional<User> findByEmail(String email);
 
-    List<User> findAllByRole(Integer roleId, int offset, int limit) throws HttpException;
+    List<User> findAllByRole(Long roleId, int offset, int limit) throws HttpException;
 
-    List<Playlist> getPlaylists(Integer userId, int offset, int limit) throws HttpException;
+    List<Playlist> getPlaylists(Long userId, int offset, int limit) throws HttpException;
 
-    List<Artist> getLikedArtists(Integer userId, int offset, int limit) throws HttpException;
+    List<Artist> getLikedArtists(Long userId, int offset, int limit) throws HttpException;
 
-    List<Album> getLikedAlbums(Integer userId, int offset, int limit) throws HttpException;
+    List<Album> getLikedAlbums(Long userId, int offset, int limit) throws HttpException;
 
-    List<Track> getLikedTracks(Integer userId, int offset, int limit) throws HttpException;
+    List<Track> getLikedTracks(Long userId, int offset, int limit) throws HttpException;
 
-    void setRole(Integer userId, Integer roleId) throws HttpException;
+    void setRole(Long userId, Integer roleId) throws HttpException;
 
-    Optional<Role> getRole(Integer userId);
+    Optional<Role> getRole(Long userId);
 
-    void addPlaylist(Integer userId, Integer playlistId) throws HttpException;
+    void addPlaylist(Long userId, Long playlistId) throws HttpException;
 
-    void removePlayList(Integer userId, Integer playlistId) throws HttpException;
+    void removePlayList(Long userId, Long playlistId) throws HttpException;
 
-    void likeArtist(Integer userId, Integer artistId) throws HttpException;
+    void likeArtist(Long userId, Long artistId) throws HttpException;
 
-    void cancelLikeArtist(Integer userId, Integer artistId) throws HttpException;
+    void cancelLikeArtist(Long userId, Long artistId) throws HttpException;
 
-    void likeAlbum(Integer userId, Integer albumId) throws HttpException;
+    void likeAlbum(Long userId, Long albumId) throws HttpException;
 
-    void cancelLikeAlbum(Integer userId, Integer albumId) throws HttpException;
+    void cancelLikeAlbum(Long userId, Long albumId) throws HttpException;
 
-    void likeTrack(Integer userId, Integer trackId) throws HttpException;
+    void likeTrack(Long userId, Long trackId) throws HttpException;
 
-    void cancelLikeTrack(Integer userId, Integer trackId) throws HttpException;
+    void cancelLikeTrack(Long userId, Long trackId) throws HttpException;
 }
