+++ src/main/java/com/epam/musicbox/repository/impl/UserRepositoryImpl.java
@@ -67,13 +67,13 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public Optional<User> findById(Integer id) {
+    public Optional<User> findById(Long id) {
         return QueryHelper.queryOne(SQL_FIND_BY_ID, userEntityBuilder, id);
     }
 
     @Override
     public void save(User user) throws HttpException {
-        Integer userId = user.getId();
+        Long userId = user.getId();
         if (userId == null) {
             QueryHelper.update(SQL_INSERT_ONE,
                     user.getLogin(),
@@ -91,7 +91,7 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public void deleteById(Integer id) throws HttpException {
+    public void deleteById(Long id) throws HttpException {
         QueryHelper.update(SQL_DELETE_BY_ID, id);
     }
 
@@ -106,77 +106,77 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public List<User> findAllByRole(Integer roleId, int offset, int limit) throws HttpException {
+    public List<User> findAllByRole(Long roleId, int offset, int limit) throws HttpException {
         return QueryHelper.queryAll(SQL_FIND_BY_ROLE, userEntityBuilder, roleId, offset, limit);
     }
 
     @Override
-    public List<Playlist> getPlaylists(Integer userId, int offset, int limit) throws HttpException {
+    public List<Playlist> getPlaylists(Long userId, int offset, int limit) throws HttpException {
         return QueryHelper.queryAll(SQL_FIND_PLAYLISTS, playlistEntityBuilder, userId, offset, limit);
     }
 
     @Override
-    public List<Artist> getLikedArtists(Integer userId, int offset, int limit) throws HttpException {
+    public List<Artist> getLikedArtists(Long userId, int offset, int limit) throws HttpException {
         return QueryHelper.queryAll(SQL_FIND_LIKED_ARTISTS, artistEntityBuilder, offset, limit);
     }
 
     @Override
-    public List<Album> getLikedAlbums(Integer userId, int offset, int limit) throws HttpException {
+    public List<Album> getLikedAlbums(Long userId, int offset, int limit) throws HttpException {
         return QueryHelper.queryAll(SQL_FIND_LIKED_ALBUMS, albumEntityBuilder, offset, limit);
     }
 
     @Override
-    public List<Track> getLikedTracks(Integer userId, int offset, int limit) throws HttpException {
+    public List<Track> getLikedTracks(Long userId, int offset, int limit) throws HttpException {
         return QueryHelper.queryAll(SQL_FIND_LIKED_TRACKS, trackEntityBuilder, offset, limit);
     }
 
     @Override
-    public void setRole(Integer userId, Integer roleId) throws HttpException {
+    public void setRole(Long userId, Integer roleId) throws HttpException {
         QueryHelper.update(SQL_SET_ROLE, userId, roleId);
     }
 
     @Override
-    public Optional<Role> getRole(Integer userId) {
+    public Optional<Role> getRole(Long userId) {
         return QueryHelper.queryOne(SQL_GET_ROLE, roleEntityBuilder, userId);
     }
 
     @Override
-    public void addPlaylist(Integer userId, Integer playlistId) throws HttpException {
+    public void addPlaylist(Long userId, Long playlistId) throws HttpException {
         QueryHelper.update(SQL_ADD_PLAYLIST, userId, playlistId);
     }
 
     @Override
-    public void removePlayList(Integer userId, Integer playlistId) throws HttpException {
+    public void removePlayList(Long userId, Long playlistId) throws HttpException {
         QueryHelper.update(SQL_REMOVE_PLAYLIST, userId, playlistId);
     }
 
     @Override
-    public void likeArtist(Integer userId, Integer artistId) throws HttpException {
+    public void likeArtist(Long userId, Long artistId) throws HttpException {
         QueryHelper.update(SQL_LIKE_ARTIST, userId, artistId);
     }
 
     @Override
-    public void cancelLikeArtist(Integer userId, Integer artistId) throws HttpException {
+    public void cancelLikeArtist(Long userId, Long artistId) throws HttpException {
         QueryHelper.update(SQL_CANCEL_LIKE_ARTIST, userId, artistId);
     }
 
     @Override
-    public void likeAlbum(Integer userId, Integer albumId) throws HttpException {
+    public void likeAlbum(Long userId, Long albumId) throws HttpException {
         QueryHelper.update(SQL_LIKE_ALBUM, userId, albumId);
     }
 
     @Override
-    public void cancelLikeAlbum(Integer userId, Integer albumId) throws HttpException {
+    public void cancelLikeAlbum(Long userId, Long albumId) throws HttpException {
         QueryHelper.update(SQL_CANCEL_LIKE_ALBUM, userId, albumId);
     }
 
     @Override
-    public void likeTrack(Integer userId, Integer trackId) throws HttpException {
+    public void likeTrack(Long userId, Long trackId) throws HttpException {
         QueryHelper.update(SQL_LIKE_TRACK, userId, trackId);
     }
 
     @Override
-    public void cancelLikeTrack(Integer userId, Integer trackId) throws HttpException {
+    public void cancelLikeTrack(Long userId, Long trackId) throws HttpException {
         QueryHelper.update(SQL_CANCEL_LIKE_TRACK, userId, trackId);
     }
 }
