+++ src/main/java/com/epam/musicbox/service/ArtistService.java
@@ -10,9 +10,9 @@ import java.util.Optional;
 public interface ArtistService extends Service<Artist> {
     Optional<Artist> findByName(String name);
 
-    List<Track> getTracks(Integer artistId, int page) throws HttpException;
+    List<Track> getTracks(Long artistId, int page) throws HttpException;
 
-    void addTrack(Integer artistId, Integer trackId) throws HttpException;
+    void addTrack(Long artistId, Long trackId) throws HttpException;
 
-    void removeTrack(Integer artistId, Integer trackId) throws HttpException;
+    void removeTrack(Long artistId, Long trackId) throws HttpException;
 }
