+++ src/main/java/com/epam/musicbox/service/impl/UserServiceImpl.java
@@ -22,7 +22,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<User> findById(Integer id) {
+    public Optional<User> findById(Long id) {
         return userRepository.findById(id);
     }
 
@@ -32,7 +32,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteById(Integer id) throws HttpException {
+    public void deleteById(Long id) throws HttpException {
         userRepository.deleteById(id);
     }
 
@@ -47,87 +47,87 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<User> findAllByRole(Integer roleId, int page) throws HttpException {
+    public List<User> findAllByRole(Long roleId, int page) throws HttpException {
         return userRepository.findAllByRole(roleId,
                 Service.getOffset(page),
                 Service.PAGE_SIZE);
     }
 
     @Override
-    public List<Playlist> getPlaylists(Integer userId, int page) throws HttpException {
+    public List<Playlist> getPlaylists(Long userId, int page) throws HttpException {
         return userRepository.getPlaylists(userId,
                 Service.getOffset(page),
                 Service.PAGE_SIZE);
     }
 
     @Override
-    public List<Artist> getLikedArtists(Integer userId, int page) throws HttpException {
+    public List<Artist> getLikedArtists(Long userId, int page) throws HttpException {
         return userRepository.getLikedArtists(userId,
                 Service.getOffset(page),
                 Service.PAGE_SIZE);
     }
 
     @Override
-    public List<Album> getLikedAlbums(Integer userId, int page) throws HttpException {
+    public List<Album> getLikedAlbums(Long userId, int page) throws HttpException {
         return userRepository.getLikedAlbums(userId,
                 Service.getOffset(page),
                 Service.PAGE_SIZE);
     }
 
     @Override
-    public List<Track> getLikedTracks(Integer userId, int page) throws HttpException {
+    public List<Track> getLikedTracks(Long userId, int page) throws HttpException {
         return userRepository.getLikedTracks(userId,
                 Service.getOffset(page),
                 Service.PAGE_SIZE);
     }
 
     @Override
-    public void setRole(Integer userId, Integer roleId) throws HttpException {
+    public void setRole(Long userId, Integer roleId) throws HttpException {
         userRepository.setRole(userId, roleId);
     }
 
     @Override
-    public Optional<Role> getRole(Integer userId) {
+    public Optional<Role> getRole(Long userId) {
         return userRepository.getRole(userId);
     }
 
     @Override
-    public void addPlaylist(Integer userId, Integer playlistId) throws HttpException {
+    public void addPlaylist(Long userId, Long playlistId) throws HttpException {
         userRepository.addPlaylist(userId, playlistId);
     }
 
     @Override
-    public void removePlayList(Integer userId, Integer playlistId) throws HttpException {
+    public void removePlayList(Long userId, Long playlistId) throws HttpException {
         userRepository.removePlayList(userId, playlistId);
     }
 
     @Override
-    public void likeArtist(Integer userId, Integer artistId) throws HttpException {
+    public void likeArtist(Long userId, Long artistId) throws HttpException {
         userRepository.likeArtist(userId, artistId);
     }
 
     @Override
-    public void cancelLikeArtist(Integer userId, Integer artistId) throws HttpException {
+    public void cancelLikeArtist(Long userId, Long artistId) throws HttpException {
         userRepository.cancelLikeArtist(userId, artistId);
     }
 
     @Override
-    public void likeAlbum(Integer userId, Integer albumId) throws HttpException {
+    public void likeAlbum(Long userId, Long albumId) throws HttpException {
         userRepository.likeAlbum(userId, albumId);
     }
 
     @Override
-    public void cancelLikeAlbum(Integer userId, Integer albumId) throws HttpException {
+    public void cancelLikeAlbum(Long userId, Long albumId) throws HttpException {
         userRepository.cancelLikeAlbum(userId, albumId);
     }
 
     @Override
-    public void likeTrack(Integer userId, Integer trackId) throws HttpException {
+    public void likeTrack(Long userId, Long trackId) throws HttpException {
         userRepository.likeTrack(userId, trackId);
     }
 
     @Override
-    public void cancelLikeTrack(Integer userId, Integer trackId) throws HttpException {
+    public void cancelLikeTrack(Long userId, Long trackId) throws HttpException {
         userRepository.cancelLikeTrack(userId, trackId);
     }
 }
